fix: Update ResizeSensor and vttCue to work in a server context during initialization
Even though OvenPlayer obviously needs to be run client-side, there can still be scenarios where it is imported server-side.
An example of this is when using the official vue component in a Nuxt app with SSR enabled.
Even when the component is only used client-side using `ClientOnly`, the import of OvenPlayer on the server has side-effects that want to access window.
By making these not error out, the Vue component can now be correctly used with Nuxt and SSR.